added express-validator and validation to genres as an example

+++ controllers/genresController.js
@@ -1,6 +1,14 @@
 const db = require("../db/queries");
+const { validationResult } = require("express-validator");
 
 const genreAddPost = async (req, res) => {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    console.log("Validation errors:", errors.array());
+    return res.redirect("/genres");
+  }
+
   console.log("Add genre to the db");
   const name = req.body.genreName;
 
@@ -24,11 +32,18 @@ const genresAllGet = async (req, res) => {
 };
 
 const genreDeletePost = async (req, res) => {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    console.log("Validation errors:", errors.array());
+    return res.redirect("/genres");
+  }
+
   console.log("Remove genre from the db");
-  const genreId = req.body.id;
+  const { genreId } = req.body;
 
   if (!genreId) {
-    return res.status(400).json({ success: false, message: "Missing genre ID."})
+    return res.status(400).json({ success: false, message: "Missing genre ID." });
   } else {
     try {
       await db.deleteGenre(genreId);
@@ -38,24 +53,26 @@ const genreDeletePost = async (req, res) => {
       res.status(500).json({ success: false, message: "There was a server error during deletion" });
     }
   }
-}
+};
 
 const genreEditPost = async (req, res) => {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    console.log("Validation errors:", errors.array());
+    return res.redirect("/genres");
+  }
+
   console.log("Edit genre :)");
   const { genreId, genreName } = req.body;
 
-  if (!genreId) {
-    res.status(400).json({success: false, message: "Missing actor ID."})
-  } else {
-    try {
-      await db.editGenre(genreId, genreName)
-      res.redirect("/genres");;
-    } catch (error) {
-      console.error(error);
-      res.redirect("/genres");;
-    }
+  try {
+    await db.editGenre(genreId, genreName);
+    res.redirect("/genres");
+  } catch (error) {
+    console.error(error);
+    res.redirect("/genres");
   }
-
 };
 
 module.exports = { genreAddPost, genresAllGet, genreDeletePost, genreEditPost };
